[Android Auto] Addressing PR feedback. Renamed SpeedLimitBitmapProvider -> SpeedLimitBitmapRenderer. Moved MutcdSpeedLimitDrawable and ViennaSpeedLimitDrawable to separate files
